Update ss58-registry.json for Polkadex
Adding prefix for Polkadex. Polkadex is a fully decentralized, peer-to-peer orderbook-based cryptocurrency exchange for the Defi ecosystem in substrate. With Polkadex, you get the same trading functionality and performance of a centralized exchange, but with the added benefits of better user-experience and fully decentralized security. Polkadex is unlocking the existing and future financial markets for everyone, bridging traditional money markets to a decentralized store of value and trust, thereby preparing the world for a tokenized future.